Переименовал солюшн и проекты, поменял нэймспэйсы, удалил конфиг файл для варнингов
