--- Enterprise/.editorconfig
@@ -1,13 +1,0 @@
-﻿[*.cs]
-
-# CS8603: Possible null reference return.
-dotnet_diagnostic.CS8603.severity = none
-
-# CS8618: Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
-dotnet_diagnostic.CS8618.severity = none
-
-# CS8602: Dereference of a possibly null reference.
-dotnet_diagnostic.CS8602.severity = none
-
-# CS8625: Cannot convert null literal to non-nullable reference type.
-dotnet_diagnostic.CS8625.severity = none
