--- Enterprise/Enterprise.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.10.35004.147
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Query.Impl", "Query.Impl\Query.Impl.csproj", "{A9893823-5B24-4340-8AD0-E07F3496E949}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "QueriesClassLibraryTests", "QueriesClassLibraryTests\QueriesClassLibraryTests.csproj", "{6C1890AB-18FE-4325-A35E-6483499D0944}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A9893823-5B24-4340-8AD0-E07F3496E949}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A9893823-5B24-4340-8AD0-E07F3496E949}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A9893823-5B24-4340-8AD0-E07F3496E949}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A9893823-5B24-4340-8AD0-E07F3496E949}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6C1890AB-18FE-4325-A35E-6483499D0944}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6C1890AB-18FE-4325-A35E-6483499D0944}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6C1890AB-18FE-4325-A35E-6483499D0944}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6C1890AB-18FE-4325-A35E-6483499D0944}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {E83F52E3-FB94-4E66-BEFB-270C0FBD2D22}
-	EndGlobalSection
-EndGlobal
