--- Enterprise/Query.Impl/Enums/AcademicRank.cs
@@ -1,18 +1,0 @@
-﻿namespace Query.Impl.Enums
-{
-    /// <summary>
-    /// Ученое звание
-    /// </summary>
-    public enum AcademicRank
-    {
-        /// <summary>
-        /// Доцент
-        /// </summary>
-        Docent = 1,
-
-        /// <summary>
-        /// Профессор
-        /// </summary>
-        Professor
-    }
-}
