--- Enterprise/Query.Impl/Enums/BuildingOwnership .cs
@@ -1,23 +1,0 @@
-﻿namespace Query.Impl.Enums
-{
-    /// <summary>
-    /// Собственность здания
-    /// </summary>
-    public enum BuildingOwnership
-    {
-        /// <summary>
-        /// Федеральная
-        /// </summary>
-        Federal = 1,
-
-        /// <summary>
-        /// Муниципальная
-        /// </summary>
-        Municipality,
-
-        /// <summary>
-        /// Частная
-        /// </summary>
-        Personal
-    }
-}
