--- Enterprise/Query.Impl/Enums/InstitutionOwnership.cs
@@ -1,18 +1,0 @@
-﻿namespace Query.Impl.Enums
-{
-    /// <summary>
-    /// Собственность учреждения
-    /// </summary>
-    public enum InstitutionOwnership
-    {
-        /// <summary>
-        /// Муниципальная
-        /// </summary>
-        Municipality = 1,
-
-        /// <summary>
-        /// Частная
-        /// </summary>
-        Personal
-    }
-}
