--- Enterprise/Query.Impl/Enums/ScientificDegree.cs
@@ -1,18 +1,0 @@
-﻿namespace Query.Impl.Enums
-{
-    /// <summary>
-    /// Ученая степень
-    /// </summary>
-    public enum ScientificDegree
-    {
-        /// <summary>
-        /// Кандидат наук
-        /// </summary>
-        Candidate = 1,
-
-        /// <summary>
-        /// Доктор наук
-        /// </summary>
-        Doctor
-    }
-}
