--- Enterprise/Query.Impl/Models/BaseModel/Entity.cs
@@ -1,31 +1,0 @@
-﻿namespace Query.Impl.Models.BaseModel
-{
-    /// <summary>
-    /// Базовый класс для сущности
-    /// </summary>
-    public abstract class Entity
-    {
-        #region ctors
-        /// <summary>
-        /// ctor
-        /// </summary>
-        protected Entity() 
-        {
-            Id = Guid.NewGuid();
-            Version = DateTime.Now;
-        }
-        #endregion
-
-        #region Fields
-        /// <summary>
-        /// 
-        /// </summary>
-        public Guid Id { get; set; }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        public DateTime Version { get; set; }
-        #endregion
-    }
-}
