+++ InstitutionStatistic/Query.Impl/Enums/AcademicRank.cs
@@ -1,0 +1,17 @@
+﻿namespace InstitutionStatistic.Query.Enums;
+
+/// <summary>
+/// Ученое звание
+/// </summary>
+public enum AcademicRank
+{
+    /// <summary>
+    /// Доцент
+    /// </summary>
+    Docent = 1,
+
+    /// <summary>
+    /// Профессор
+    /// </summary>
+    Professor
+}
