+++ InstitutionStatistic/Query.Impl/Enums/BuildingOwnership .cs
@@ -1,0 +1,22 @@
+﻿namespace InstitutionStatistic.Query.Enums;
+
+/// <summary>
+/// Собственность здания
+/// </summary>
+public enum BuildingOwnership
+{
+    /// <summary>
+    /// Федеральная
+    /// </summary>
+    Federal = 1,
+
+    /// <summary>
+    /// Муниципальная
+    /// </summary>
+    Municipality,
+
+    /// <summary>
+    /// Частная
+    /// </summary>
+    Personal
+}
