+++ InstitutionStatistic/Query.Impl/Enums/InstitutionOwnership.cs
@@ -1,0 +1,18 @@
+﻿namespace InstitutionStatistic.Query.Enums;
+
+/// <summary>
+/// Собственность учреждения
+/// </summary>
+public enum InstitutionOwnership
+    {
+        /// <summary>
+        /// Муниципальная
+        /// </summary>
+        Municipality = 1,
+
+        /// <summary>
+        /// Частная
+        /// </summary>
+        Personal
+    }
+
