+++ InstitutionStatistic/Query.Impl/Enums/ScientificDegree.cs
@@ -1,0 +1,17 @@
+﻿namespace InstitutionStatistic.Query.Enums;
+
+/// <summary>
+/// Ученая степень
+/// </summary>
+public enum ScientificDegree
+{
+    /// <summary>
+    /// Кандидат наук
+    /// </summary>
+    Candidate = 1,
+
+    /// <summary>
+    /// Доктор наук
+    /// </summary>
+    Doctor
+}
