+++ InstitutionStatistic/Query.Impl/Queries/SpecialityQuery.cs
@@ -1,0 +1,28 @@
+﻿using InstitutionStatistic.Query.Models;
+
+namespace InstitutionStatistic.Query.Queries;
+
+
+/// <summary>
+/// Запросы о специальностях
+/// </summary>
+public class SpecialityQuery : GetInfoQuery<Speciality>
+{
+    /// <summary>
+    /// ctor
+    /// </summary>
+    /// <param name="repository"></param>
+    public SpecialityQuery(ICollection<Speciality> repository) : base(repository) { }
+
+    /// <summary>
+    /// Получить топ 5 специальностей
+    /// </summary>
+    /// <returns></returns>
+    public List<Speciality> GetTopFiveSpecialities()
+    {
+        return Repository
+            .OrderByDescending(x => x.Groups.Count)
+            .Take(5)
+            .ToList();
+    }
+}
